minor refactors/cleanup
