--- src/components/QuickFIlterInput/QuickFilterInput.tsx
@@ -1,13 +1,0 @@
-import { TextInput, type TextInputProps } from "@mantine/core"
-
-export const QuickFilterInput = (props: TextInputProps) => {
-	return (
-		<TextInput
-			styles={{ label: { color: "white" } }}
-			label="Quick Filter"
-			placeholder="Search the entire table..."
-			w={400}
-			{...props}
-		/>
-	)
-}
