+++ src/features/ReminderTable/ReminderTable.tsx
@@ -8,22 +8,42 @@ import { Button, Center, Group, Stack, Tooltip } from "@mantine/core"
 
 import styles from "./ReminderTable.module.css"
 import { type ReminderData } from "../../types/types"
-import { useInput } from "../../hooks/useInput"
+
 import { columnDefs, defaultColDef } from "./columnDefs"
 import { withReminderFormProvider } from "../../providers/ReminderFormProvider"
 import { useDebouncedValue } from "@mantine/hooks"
-import { QuickFilterInput } from "../../components/QuickFIlterInput/QuickFilterInput"
+
 import { ReminderTableModal } from "./ReminderTableModal"
 import { observer } from "mobx-react"
 import { useOpenReminderTableCreateModal } from "../../hooks/reminderTable"
 import { ExportDataButton, jsonDataProps } from "../../components/buttons/ExportDataButton"
+import { QuickFilterInput, useQuickFilterInput } from "../../components/input/QuickFilterInput"
 
 type ReminderTableProps = {
 	data: ReminderData[]
 }
 
 const _ReminderTable = ({ data }: ReminderTableProps) => {
-	const { gridRef, onGridReady, resetColumns, inputProps, debouncedInputValue } = useTable()
+	const gridRef = useRef<AgGridReact<ReminderData>>(null as unknown as AgGridReact<ReminderData>) // assuming it is used only in a handler
+
+	const { value, inputProps, setValue } = useQuickFilterInput()
+	const [debouncedValue] = useDebouncedValue(value, 250)
+
+	const autoSizeStaticColumns = useCallback(() => {
+		gridRef.current.columnApi.autoSizeColumns(["id", "actions"])
+	}, [])
+
+	const resetColumns = useCallback(() => {
+		gridRef.current.columnApi.resetColumnState()
+		autoSizeStaticColumns()
+		gridRef.current.api.setFilterModel(null)
+		setValue("")
+	}, [autoSizeStaticColumns, setValue])
+
+	const onGridReady = useCallback(() => {
+		autoSizeStaticColumns()
+	}, [autoSizeStaticColumns])
+
 	const openNew = useOpenReminderTableCreateModal()
 
 	return (
@@ -50,7 +70,7 @@ const _ReminderTable = ({ data }: ReminderTableProps) => {
 							ref={gridRef}
 							pagination={true}
 							ensureDomOrder={true}
-							quickFilterText={debouncedInputValue}
+							quickFilterText={debouncedValue}
 							rowData={data}
 							tooltipShowDelay={400}
 							columnDefs={columnDefs}
@@ -66,34 +86,3 @@ const _ReminderTable = ({ data }: ReminderTableProps) => {
 }
 
 export const ReminderTable = withReminderFormProvider(observer(_ReminderTable))
-
-const useTable = () => {
-	const gridRef = useRef<AgGridReact<ReminderData>>(null as any as AgGridReact<ReminderData>) // assuming it is used only in a handler
-
-	const { value, inputProps, setValue } = useInput()
-	const [debouncedValue] = useDebouncedValue(value, 250)
-
-	const autoSizeStaticColumns = useCallback(() => {
-		gridRef.current.columnApi.autoSizeColumns(["id", "actions"])
-	}, [])
-
-	const resetColumns = useCallback(() => {
-		gridRef.current.columnApi.resetColumnState()
-		autoSizeStaticColumns()
-		gridRef.current.api.setFilterModel(null)
-		setValue("")
-	}, [autoSizeStaticColumns, setValue])
-
-	const onGridReady = useCallback(() => {
-		autoSizeStaticColumns()
-	}, [autoSizeStaticColumns])
-
-	return {
-		gridRef,
-		inputProps,
-		debouncedInputValue: debouncedValue,
-		resetColumns,
-		autoSizeStaticColumns,
-		onGridReady,
-	}
-}
